refactor: Optimize WxOpenController and update versions
Refactored WxOpenController.cs to reduce redundant calls by using
a local variable `weixinSetting`. Updated logic for merchant ID,
notification URL, and other parameters to use `weixinSetting`.
Adjusted `GetJsApiUiPackage` and result `appId` to align with
`weixinSetting`.

Updated project version numbers across multiple `.csproj` files,
including Senparc.Weixin and its modules, to reflect minor and
patch-level changes. Notable updates include support for Work
module refactoring in Senparc.Weixin.All.csproj.

Fixed `MsgID` property in RequestMessageEvent_MassSendJobFinish.cs
to ensure it returns a `long` type.

+++ src/Senparc.Weixin.MP/Senparc.Weixin.MP/Entities/Request/Event/RequestMessageEvent_MassSendJobFinish.cs
@@ -140,7 +140,7 @@ namespace Senparc.Weixin.MP.Entities
         /// </summary>
         public long MsgID
         {
-            get { return base.MsgId; }
+            get { return (long)base.MsgId; }
             set
             {
                 base.MsgId = value;
